move dictionary array specific code to array/dict_internal.{h,cc}
