+++ cpp/src/arrow/array.cc
@@ -37,7 +37,6 @@
 #include "arrow/util/bit_util.h"
 #include "arrow/util/checked_cast.h"
 #include "arrow/util/decimal.h"
-#include "arrow/util/int_util.h"
 #include "arrow/util/logging.h"
 #include "arrow/util/macros.h"
 #include "arrow/visitor.h"
@@ -911,124 +910,6 @@ Status DictionaryArray::FromArrays(const std::shared_ptr<DataType>& type,
   RETURN_NOT_OK(is_valid);
   *out = std::make_shared<DictionaryArray>(type, indices, dictionary);
   return Status::OK();
-}
-
-static bool IsTrivialTransposition(const std::vector<int32_t>& transpose_map,
-                                   int64_t input_dict_size) {
-  for (int64_t i = 0; i < input_dict_size; ++i) {
-    if (transpose_map[i] != i) {
-      return false;
-    }
-  }
-  return true;
-}
-
-template <typename InType, typename OutType>
-static Status TransposeDictIndices(MemoryPool* pool, const ArrayData& in_data,
-                                   const std::vector<int32_t>& transpose_map,
-                                   const std::shared_ptr<ArrayData>& out_data,
-                                   std::shared_ptr<Array>* out) {
-  using in_c_type = typename InType::c_type;
-  using out_c_type = typename OutType::c_type;
-  internal::TransposeInts(in_data.GetValues<in_c_type>(1),
-                          out_data->GetMutableValues<out_c_type>(1), in_data.length,
-                          transpose_map.data());
-  *out = MakeArray(out_data);
-  return Status::OK();
-}
-
-Status DictionaryArray::Transpose(MemoryPool* pool, const std::shared_ptr<DataType>& type,
-                                  const std::shared_ptr<Array>& dictionary,
-                                  const std::vector<int32_t>& transpose_map,
-                                  std::shared_ptr<Array>* out) const {
-  if (type->id() != Type::DICTIONARY) {
-    return Status::TypeError("Expected dictionary type");
-  }
-  const int64_t in_dict_len = data_->dictionary->length();
-  if (in_dict_len > static_cast<int64_t>(transpose_map.size())) {
-    return Status::Invalid(
-        "Transpose map too small for dictionary array "
-        "(has ",
-        transpose_map.size(), " values, need at least ", in_dict_len, ")");
-  }
-
-  const auto& out_dict_type = checked_cast<const DictionaryType&>(*type);
-
-  const auto& out_index_type =
-      static_cast<const FixedWidthType&>(*out_dict_type.index_type());
-
-  auto in_type_id = dict_type_->index_type()->id();
-  auto out_type_id = out_index_type.id();
-
-  if (in_type_id == out_type_id && IsTrivialTransposition(transpose_map, in_dict_len)) {
-    // Index type and values will be identical => we can simply reuse
-    // the existing buffers.
-    auto out_data =
-        ArrayData::Make(type, data_->length, {data_->buffers[0], data_->buffers[1]},
-                        data_->null_count, data_->offset);
-    out_data->dictionary = dictionary;
-    *out = MakeArray(out_data);
-    return Status::OK();
-  }
-
-  // Default path: compute a buffer of transposed indices.
-  std::shared_ptr<Buffer> out_buffer;
-  RETURN_NOT_OK(AllocateBuffer(
-      pool, data_->length * out_index_type.bit_width() * CHAR_BIT, &out_buffer));
-
-  // Shift null buffer if the original offset is non-zero
-  std::shared_ptr<Buffer> null_bitmap;
-  if (data_->offset != 0) {
-    RETURN_NOT_OK(
-        CopyBitmap(pool, null_bitmap_data_, data_->offset, data_->length, &null_bitmap));
-  } else {
-    null_bitmap = data_->buffers[0];
-  }
-
-  auto out_data =
-      ArrayData::Make(type, data_->length, {null_bitmap, out_buffer}, data_->null_count);
-  out_data->dictionary = dictionary;
-
-#define TRANSPOSE_IN_OUT_CASE(IN_INDEX_TYPE, OUT_INDEX_TYPE)    \
-  case OUT_INDEX_TYPE::type_id:                                 \
-    return TransposeDictIndices<IN_INDEX_TYPE, OUT_INDEX_TYPE>( \
-        pool, *data_, transpose_map, out_data, out);
-
-#define TRANSPOSE_IN_CASE(IN_INDEX_TYPE)                        \
-  case IN_INDEX_TYPE::type_id:                                  \
-    switch (out_type_id) {                                      \
-      TRANSPOSE_IN_OUT_CASE(IN_INDEX_TYPE, Int8Type)            \
-      TRANSPOSE_IN_OUT_CASE(IN_INDEX_TYPE, Int16Type)           \
-      TRANSPOSE_IN_OUT_CASE(IN_INDEX_TYPE, Int32Type)           \
-      TRANSPOSE_IN_OUT_CASE(IN_INDEX_TYPE, Int64Type)           \
-      default:                                                  \
-        return Status::NotImplemented("unexpected index type"); \
-    }
-
-  switch (in_type_id) {
-    TRANSPOSE_IN_CASE(Int8Type)
-    TRANSPOSE_IN_CASE(Int16Type)
-    TRANSPOSE_IN_CASE(Int32Type)
-    TRANSPOSE_IN_CASE(Int64Type)
-    default:
-      return Status::NotImplemented("unexpected index type");
-  }
-#undef TRANSPOSE_IN_CASE
-#undef TRANSPOSE_IN_OUT_CASE
-}
-
-Result<bool> DictionaryArray::CanCompareIndices(const DictionaryArray& other) const {
-  if (!dictionary()->type()->Equals(other.dictionary()->type())) {
-    return Status::TypeError("dictionaries have differing type ", *dictionary()->type(),
-                             " vs ", *other.dictionary()->type());
-  }
-
-  if (!indices()->type()->Equals(other.indices()->type())) {
-    return false;
-  }
-
-  auto min_length = std::min(dictionary()->length(), other.dictionary()->length());
-  return dictionary()->RangeEquals(other.dictionary(), 0, min_length, 0);
 }
 
 // ----------------------------------------------------------------------
